+++ integration/sql/impl/src/visitor.rs
@@ -1,20 +1,18 @@
 // Copyright 2018-2023 contributors to the OpenLineage project
 // SPDX-License-Identifier: Apache-2.0
 
-use core::time;
-use std::sync::Arc;
-
 use crate::context::Context;
 use crate::lineage::*;
 
 use anyhow::{anyhow, Result};
-use sqlparser::ast::{AlterTableOperation, Expr, Function, FunctionArg, FunctionArgExpr, Ident, ListAggOnOverflow, OrderByExpr, Query, Select, SelectItem, SetExpr, Statement, Table, TableAlias, TableFactor, TableWithJoins, WindowSpec, With};
-use sqlparser::test_utils::table;
+use sqlparser::ast::{
+    AlterTableOperation, Expr, Function, FunctionArg, FunctionArgExpr, ListAggOnOverflow, Query,
+    Select, SelectItem, SetExpr, Statement, Table, TableFactor, WindowSpec, With,
+};
 
 pub trait Visit {
     fn visit(&self, context: &mut Context) -> Result<()>;
 }
-
 impl Visit for With {
     fn visit(&self, context: &mut Context) -> Result<()> {
         for cte in &self.cte_tables {
@@ -52,7 +50,19 @@ impl Visit for TableFactor {
                 }
                 context.add_input(name.to_string());
                 Ok(())
-            }
+            },
+            TableFactor::Pivot { name, pivot_alias, ..} => {
+                let table = DbTableMeta::new(
+                    name.to_string(),
+                    context.dialect(),
+                    context.default_schema().clone(),
+                );
+                if let Some(pivot_alias) = pivot_alias {
+                    context.add_table_alias(table.clone(), pivot_alias.name.value.clone());
+                }
+                context.add_input(name.to_string());
+                Ok(())
+            },
             TableFactor::Derived {
                 lateral: _,
                 subquery,
@@ -153,7 +163,7 @@ impl Visit for Expr {
                     let table = DbTableMeta::new(
                         prefix,
                         context.dialect(),
-                        context.default_schema().clone(),
+                        context.default_schema().clone()
                     );
                     context.add_column_ancestors(
                         ColumnMeta::new(descendant, None),
@@ -394,7 +404,7 @@ impl Visit for Select {
                 let table = DbTableMeta::new(
                     name.to_string(),
                     context.dialect(),
-                    context.default_schema().clone(),
+                    context.default_schema().clone()
                 );
                 if let Some(alias) = alias {
                     context.add_table_alias(table.clone(), alias.name.value.clone());
@@ -475,7 +485,7 @@ impl Visit for SetExpr {
                 left.visit(context)?;
                 right.visit(context)
             }
-            SetExpr::Table(table) => table.visit(context)
+            SetExpr::Table(table) => table.visit(context),
         }
     }
 }
@@ -506,7 +516,6 @@ impl Visit for Query {
 impl Visit for Statement {
     fn visit(&self, context: &mut Context) -> Result<()> {
         context.push_frame();
-
         match self {
             Statement::Query(query) => query.visit(context)?,
             Statement::Insert {
@@ -539,12 +548,30 @@ impl Visit for Statement {
 
                 context.add_output(name.to_string());
             }
+            Statement::CreateView { name, query, .. } => {
+                query.visit(context)?;
+                context.add_output(name.to_string());
+            },
+            Statement::CreateStage { name, stage_params, .. } => {
+                if stage_params.url.as_ref().is_some() {
+                    context.add_non_table_input(
+                        stage_params.url.as_ref().unwrap().to_string(),
+                        true,
+                        true
+                    );
+                }
+                context.add_non_table_output(
+                    name.to_string(),
+                    false,
+                    true
+                );
+            },
             Statement::Update {
                 table,
                 assignments: _,
                 from,
                 selection,
-                returning
+                ..
             } => {
                 let name = get_table_name_from_table_factor(&table.relation)?;
                 context.add_output(name);
@@ -559,11 +586,7 @@ impl Visit for Statement {
                     expr.visit(context)?;
                 }
             }
-            Statement::AlterTable {
-                name,
-                operation
-            } => {
-
+            Statement::AlterTable { name, operation } => {
                 match operation {
                     AlterTableOperation::SwapWith { table_name } => {
                         // both table names are inputs and outputs of the swap operation
@@ -577,16 +600,14 @@ impl Visit for Statement {
                         context.add_input(name.to_string());
                         context.add_output(table_name.to_string());
                     }
-                    _ => {
-                        context.add_output(name.to_string())
-                    }
+                    _ => context.add_output(name.to_string()),
                 }
             }
             Statement::Delete {
                 table_name,
                 using,
                 selection,
-                returning
+                ..
             } => {
                 let table_name = get_table_name_from_table_factor(table_name)?;
                 context.add_output(table_name);
@@ -599,23 +620,31 @@ impl Visit for Statement {
                     expr.visit(context)?;
                 }
             }
-            Statement::Truncate {
-                table_name,
-                partitions
-            } => {
-                context.add_output(table_name.to_string())
-            }
-            Statement::Drop {
-                object_type,
-                if_exists,
-                names,
-                cascade,
-                restrict,
-                purge
-            } => {
+            Statement::Truncate { table_name, .. } => context.add_output(table_name.to_string()),
+            Statement::Drop { names, .. } => {
                 for name in names {
                     context.add_output(name.to_string())
                 }
+            }
+            Statement::CopyIntoSnowflake {
+                into, from_stage, ..
+            } => {
+                context.add_output(into.to_string());
+                if from_stage.to_string().contains("gcs://") ||
+                    from_stage.to_string().contains("s3://") ||
+                    from_stage.to_string().contains("azure://") {
+                    context.add_non_table_input(
+                        from_stage
+                            .to_string()
+                            .replace("'", "")
+                            .replace("\"", ""), // just unquoted location URL with,
+                        true,
+                        true
+                    );
+                } else {
+                    // Stage
+                    context.add_non_table_input(from_stage.to_string(), true, true);
+                };
             }
             _ => {}
         }